requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.24.0
openai==1.84.*
pandas==2.1.4
geopandas==0.14.1
python-multipart==0.0.6
aiofiles==23.2.1
jinja2==3.1.2
httpx==0.28.*
numpy==1.25.2
shapely==2.0.2
python-dotenv==1.0.0
pillow==10.1.0
pydantic==2.7.*
# starlette==0.43.*
fastapi-mcp==0.3.* 